# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "System manage music",
        "title": "Music API REST",
        "termsOfService": "https://juancx.site/",
        "contact": {
            "name": "API Support",
            "email": "juanm.campos@unas.edu.pe"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:4021",
    "basePath": "/",
    "paths": {
...
        "/api/v1/songs-play-list": {
            "put": {
                "description": "Método que permite Actualiza una instancia del objeto SongsOfPlaylist en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsOfPlaylist"
                ],
                "summary": "Actualiza una instancia de SongsOfPlaylist",
                "parameters": [
                    {
                        "description": "Datos para actualizar SongsOfPlaylist",
                        "name": "RequestSongsOfPlaylist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.RequestSongsOfPlaylist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    }
                }
            },
            "post": {
                "description": "Método que permite crear una instancia del objeto SongsOfPlaylist en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsOfPlaylist"
                ],
                "summary": "Crear una instancia de SongsOfPlaylist",
                "parameters": [
                    {
                        "description": "Datos para crear SongsOfPlaylist",
                        "name": "RequestSongsOfPlaylist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.RequestSongsOfPlaylist"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    }
                }
            }
        },
        "/api/v1/songs-play-list/": {
            "get": {
                "description": "Método que permite obtener todas las instancias del objeto SongsOfPlaylist en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsOfPlaylist"
                ],
                "summary": "Obtiene todas las instancias de SongsOfPlaylist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    }
                }
            }
        },
        "/api/v1/songs-play-list/{id}": {
            "get": {
                "description": "Método que permite obtener una instancia del objeto SongsOfPlaylist en la base de datos por su id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsOfPlaylist"
                ],
                "summary": "Obtiene una instancia de SongsOfPlaylist por su id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SongsOfPlaylist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    }
                }
            },
            "delete": {
                "description": "Método que permite eliminar una instancia del objeto SongsOfPlaylist en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsOfPlaylist"
                ],
                "summary": "Elimina una instancia de SongsOfPlaylist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SongsOfPlaylist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_play_list.ResponseSongsOfPlaylist"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "songs_play_list.RequestSongsOfPlaylist": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "playlist": {
                    "type": "integer"
                },
                "song": {
                    "type": "integer"
                }
            }
        },
        "songs_play_list.ResponseSongsOfPlaylist": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}